Added UnitTests to the project to test the authorization of the user. Because of the test project, the IdentityServerConfig project has been moved to the /src folder
