+++ src/IdentityServerConfig.csproj
