+++ src/Pages/ApiResourceDetails.razor
