+++ src/Pages/ClientDetails.razor
