+++ src/Pages/ReferenceToken.razor
@@ -1,5 +1,5 @@
 ﻿@page "/reference-token"
-@attribute [Authorize(Policy = "IDSCReferenceTokenView")]
+@attribute [Authorize(Policy = "ReferenceTokenView")]
 @inject PersistedGrantDbContext DbContext
 
 <ul class="nav nav-tabs mb-3" id="referenceTokenTabs" role="tablist">
