+++ src/Pages/ScopeDetails.razor
