+++ src/appsettings.Development.json
